+++ src/ext/bio.rs
@@ -1,7 +1,7 @@
 use crate::binary::RankerT;
 
 impl RankerT for bio::data_structures::rank_select::RankSelect {
-    fn new_packed(seq: &[usize]) -> Self {
+    fn new_packed(seq: &[u64]) -> Self {
         let seq_u8 = unsafe { seq.align_to::<u8>().1 };
         let bitvec = bv::BitVec::<u8>::from_bits(seq_u8);
         let k = 512 / 32;
